Change dependency from mokutil-git to mokutil
The reason we chose mokutil-git over mokutil was a build issue for the
mokutil package, whereas mokutil-git did work. Unfortunately,
mokutil-git lacks the provides and replaces field, so this had to be set
explicitly in our dependencies. This issue has now been resolved, so
switch to the hopefully more stable non-git version.